Fix expected column type for some select4.test variations
The expected column types for a number of variations in select4.test
appears to be incorrect. These tests should expect Integer, not Text.
I suspect the original SqlLite C test harness is lax about type checking,
and accepts integers as text, and therefore doesn't error. But since
the Cockroach harness is more strict, it's necessary to change the
expected type to exactly match the actual type returned by the query.